Major Changes
-----------------------------------------------
More work on the inbox controller. You can now subscribe to categories feeds.


git-svn-id: svn+ssh://svn.locusfoc.us/home/svn/ram/trunk@760 7bd17f4f-1217-0410-bbc5-bc4367116c30

+++ app/models/group.rb
@@ -1,4 +1,4 @@
-# Schema as of Sun Oct 22 21:28:20 PDT 2006 (schema version 19)
+# Schema as of Fri Oct 27 20:31:51 PDT 2006 (schema version 22)
 #
 #  id                  :integer(11)   not null
 #  name                :string(255)   
@@ -10,105 +10,120 @@
 #  parent_id           :integer(11)   
 #  counter_cache       :boolean(1)    default(true)
 #  permanent           :boolean(1)    
+#  created_at          :datetime      
+#  updated_at          :datetime      
+#
+
+# Schema as of Fri Oct 27 16:48:28 PDT 2006 (schema version 21)
+#
+#  id                  :integer(11)   not null
+#  name                :string(255)   
+#  description         :text          
+#  public              :boolean(1)    default(true)
+#  user_id             :integer(11)   
+#  type                :string(255)   
+#  state_id            :integer(11)   
+#  parent_id           :integer(11)   
+#  counter_cache       :boolean(1)    default(true)
+#  permanent           :boolean(1)    
+#  created_at          :datetime      
+#  updated_at          :datetime      
 #
 
 class Group < Collection
   acts_as_taggable
   include TagMethods
 
-  has_many :memberships, :foreign_key=>:collection_id
-  has_many :users, :through => :memberships, :conditions => "memberships.collection_type = 'Group'" do
-    def <<(user)
-      return if @owner.users.include?user
+	has_many :memberships, :foreign_key=>:collection_id
+	has_many :users, :through => :memberships, :conditions => "memberships.collection_type = 'Group'" do
+	  def <<(user)
+	    return if @owner.users.include?user
+	    m = Membership.create(
+        :user_id => user.id,
+        :collection_id => @owner.id,
+        :collection_type => 'Group' #@owner.class.class_name	    
+	    )
+	    m.save!
+		
+      # After we add a user to a group, we want the associated 'users' object to have the current list of users automatically,
+      # so we must reload the users cache here. If this behavior is not desired, comment out this line.
+      @owner.users(true) 
 
-      @owner.transaction do
-        m = Membership.create(
-          :user_id => user.id,
-          :collection_id => @owner.id,
-          :collection_type => 'Group' #@owner.class.class_name      
-        )
-        GroupObserver.after_add(@owner, user)
-    
-        # After we add a user to a group, we want the associated 'users' object to have the current list of users automatically,
-        # so we must reload the users cache here. If this behavior is not desired, comment out this line.
-        @owner.users(true) 
-      end
-
-      #TODO: add category ids to user's category tree
+		  #TODO: add category ids to user's category tree
     end
-  end
-  
-  has_many :linkings
-  
-  #TODO: Possibly make this more generic and accept many polymorphic types by using @owner.class.class_name in place of assets
-  has_many :assets, :through=> :linkings, :source=>:asset, :conditions=>"linkings.linkable_type='Asset'" do
-    def <<(asset)
-      return false if @owner.assets.include?asset
-      l = Linking.create(
-        :linkable_id => asset.id,
-        :linkable_type => 'Asset',
-        :group_id => @owner.id
-      )
-      l.save!
-    end
-  end
-  
-  has_many :articles, :through=> :linkings, :source=>:article, :conditions=>"linkings.linkable_type='Article'" do
-    def <<(article)
-      return false if @owner.articles.include?article
-      l = Linking.create(
-        :linkable_id => article.id,
-        :linkable_type => 'Article',
-        :group_id => @owner.id
-      )
-      l.save!
-    end
-  end
-  
-  has_many :changes, :finder_sql=>'SELECT DISTINCT * ' +
+	end
+	
+	has_many :linkings
+	
+	#TODO: Possibly make this more generic and accept many polymorphic types by using @owner.class.class_name in place of assets
+	has_many :assets, :through=> :linkings, :source=>:asset, :conditions=>"linkings.linkable_type='Asset'" do
+	  def <<(asset)
+	    return false if @owner.assets.include?asset
+	    l = Linking.create(
+	      :linkable_id => asset.id,
+	      :linkable_type => 'Asset',
+	      :group_id => @owner.id
+	    )
+	    l.save!
+	  end
+	end
+	
+	has_many :articles, :through=> :linkings, :source=>:article, :conditions=>"linkings.linkable_type='Article'" do
+	  def <<(article)
+	    return false if @owner.articles.include?article
+	    l = Linking.create(
+	      :linkable_id => article.id,
+	      :linkable_type => 'Article',
+	      :group_id => @owner.id
+	    )
+	    l.save!
+	  end
+	end
+	
+	has_many :changes, :finder_sql=>'SELECT DISTINCT * ' +
         'FROM changes c WHERE c.record_id = #{id} AND c.record_type = "Group" ORDER BY c.created_at'
-  
-  
-  #has_many :categories, :through => :access_contexts, :foreign_key=>:category_id do
-  has_many :categories, :through => :linkings, :select => "DISTINCT collections.*", :foreign_key=>:category_id do
-    def <<(category)
-      return if @owner.categories.include?category
+	
+	
+	#has_many :categories, :through => :access_contexts, :foreign_key=>:category_id do
+	has_many :categories, :through => :linkings, :select => "DISTINCT collections.*", :foreign_key=>:category_id do
+	  def <<(category)
+	    return if @owner.categories.include?category
       a = Linking.create(
         :category_id => category.id,
         :group_id => @owner.id
-      )
-      a.save!
+	    )
+	    a.save!
     end
-  end
-  
-  def members
+	end
+	
+	def members
     users
   end
   
   def permanent?
-    permanent
+    self.permanent
   end
   
   def public?
     self.public
   end
-  
-  def leader
-    User.find(@owner.user_id)
-  end
+	
+	def leader
+		User.find(@owner.user_id)
+	end
   
   # switch_ownership(5)
   def switch_ownership(new_owner)
-    old_owner = user_id
+    old_owner = self.user_id
     user_id = new_owner
     save!
-    m = Membership.create(:user_id => old_owner, :state_id => 2, :collection_id => id, :collection_type => 'Group')    
+    m = Membership.create(:user_id => old_owner, :state_id => 2, :collection_id => self.id, :collection_type => 'Group')    
     memberships.find_by_user_id(new_owner).destroy rescue nil
   end
   
   # enemies.remove_all_members
   def remove_all_members
-    members.each do |m| 
+    self.members.each do| m | 
       remove_member(m)
     end
   end
@@ -119,25 +134,21 @@ class Group < Collection
   
   #this will also remove all assets and articles, since access ot these models is dependent on the category as well.
   def remove_all_categories
-    linkings.each do |link| 
+    self.linkings.each do| link | 
       #borked.
       link.destroy
     end
   end
   
   def remove_category(category)
-    linking = Linking.find_by_group_id_and_category_id(id, category.id)
+    linking =Linking.find_by_group_id_and_category_id(self.id, category.id)
     linking.destroy if linking.valid?
   end
   
-  # enemies.remove_member
+  # enemeies.remove_member
   def remove_member(member)
-    membership = Membership.find_by_user_id_and_collection_id(member.id, id)
-    if membership
-      membership.destroy
-      GroupObserver.after_remove(self, member)
-      true
-    end
+    membership = Membership.find_by_user_id_and_collection_id(member.id, self.id)
+    membership.destroy if membership     
   end
 
   def remove_user(user)
@@ -146,7 +157,8 @@ class Group < Collection
 
   validates_presence_of :user_id, :name
   validates_uniqueness_of :name
-  validates_unchangeable :name, { :message => "A perminant Group's name cannot be changed." , :if => Proc.new { |group| group.permanent } }
+  validates_unchangeable :name, { :message => "A perminant Group's name cannot be changed." , :if=> Proc.new{|group| group.permanent} }
+  
   
   #todo after save add user that created group to membership list
